Split the limit request issue forms and add more fields
This will make it easier to triage these tickets, by pushing users to
include relevant information directly in the issue's title and text.

This separation also makes it clearer that there's two separate limits
and that the user needs to provide require different kinds information
to help handle them quickly.